transcode-node/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
pyyaml>=6.0
psutil>=5.9.0
